requirements.txt

flask
scikit-learn
nltk
PyPDF2
python-docx
